requirements.txt
shap>=0.42.1
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
jinja2>=3.1.0
weasyprint>=57.0
openai>=1.0.0
python-dotenv>=0.19.0